requirements.txt
hydra-core
pytorch-lightning
transformers
easydict
colorama
fastnlp
nltk
wandb
matplotlib
seaborn
